requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
backports.zoneinfo==0.2.1
branca==0.4.2
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.0.14
django-apscheduler==0.6.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-import-export==2.7.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
firebase-admin==5.2.0
folium==0.12.1.post1
google-api-core==2.5.0
google-api-python-client==2.36.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.2.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
Jinja2==3.0.3
MarkupPy==1.14
MarkupSafe==2.0.1
msgpack==1.0.3
numpy==1.22.2
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
proto-plus==1.20.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyparsing==3.0.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.8
xlrd==2.0.1
xlwt==1.3.0
